requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
waitress==1.3.0
whitenoise==4.1.2
